SONARJAVA-3307 Migrate to SonarJava 6.1 and LTS 7.9
