+++ java-custom-rules/pom.xml
@@ -16,8 +16,8 @@
 		<sslr.version>1.21</sslr.version>
 		<gson.version>2.6.2</gson.version>
 
-		<sonar.version>7.7</sonar.version>
-		<sonarjava.version>5.12.1.17771</sonarjava.version>
+		<sonar.version>8.2.0.32929</sonar.version>
+		<sonarjava.version>6.1.0.20866</sonarjava.version>
 	</properties>
 
 	<dependencies>
@@ -34,6 +34,13 @@
 			<type>sonar-plugin</type>
 			<version>${sonarjava.version}</version>
 			<scope>provided</scope>
+		</dependency>
+
+		<!-- Required since SonarQube > 8.1 -->
+		<dependency>
+			<groupId>org.sonarsource.sonarqube</groupId>
+			<artifactId>sonar-plugin-api-impl</artifactId>
+			<version>${sonar.version}</version>
 		</dependency>
 
 		<dependency>
@@ -93,7 +100,7 @@
 		<dependency>
 			<groupId>com.google.guava</groupId>
 			<artifactId>guava</artifactId>
-			<version>19.0</version>
+			<version>26.0-jre</version>
 		</dependency>
 
 		<dependency>
@@ -134,7 +141,7 @@
 					<pluginName>Java Custom Rules</pluginName>
 					<pluginClass>org.sonar.samples.java.MyJavaRulesPlugin</pluginClass>
 					<sonarLintSupported>true</sonarLintSupported>
-					<sonarQubeMinVersion>6.7</sonarQubeMinVersion>
+					<sonarQubeMinVersion>7.9</sonarQubeMinVersion>
 					<requirePlugins>java:${sonarjava.version}</requirePlugins>
 				</configuration>
 			</plugin>
