+++ java-custom-rules/src/main/resources/org/sonar/l10n/java/rules/squid/AvoidTreeList_java.html
@@ -1,4 +1,4 @@
-<p>This rule detects instanciation of apache commons collections UnmodifableList"</p>
+<p>This rule detects instanciation of apache commons collections TreeList"</p>
 <h2>Noncompliant Code Example</h2>
 <pre>
 TO DO
