+++ java-custom-rules/src/main/resources/org/sonar/l10n/java/rules/squid/AvoidTreeList_java.json
@@ -1,5 +1,5 @@
 {
-  "title": "Avoid usage of UnmodifiableList",
+  "title": "Avoid usage of TreeList",
   "type": "CODE_SMELL",
   "status": "ready",
   "remediation": {
